fix editingShortcutIndex

+++ src/Engine/include/OpenLoco/Engine/Input/ShortcutManager.h
@@ -13,7 +13,7 @@ namespace OpenLoco
 
 namespace OpenLoco::Input
 {
-    enum class Shortcut : uint32_t;
+    enum class Shortcut : uint16_t;
 }
 
 namespace OpenLoco::Input::ShortcutManager
